changes -> commands
